refactor(plugin-js-packages): avoid implementation in index.ts file
